Season 3 audio update

+++ audio/audio_uploader.py
@@ -14,12 +14,14 @@ from utils.upload_utils import upload_file
 from utils.wiki_utils import s
 
 
-def upload_audio(source: Path, target: FilePage, text: str, force: bool = False):
+def upload_audio(source: Path, target: FilePage, text: str, force: bool = False, temp_wiki_file: Path = None):
     assert source.exists()
     temp_file = Path("temp.ogg")
     wav_to_ogg(source, temp_file)
     upload_file(text=text, target=target, file=temp_file, force=force)
-    temp_file.unlink()
+    if temp_wiki_file.exists():
+        temp_wiki_file.unlink()
+    temp_file.rename(temp_wiki_file)
 
 
 def upload_audio_file(voices: list[Voice],
@@ -40,28 +42,31 @@ def upload_audio_file(voices: list[Voice],
                 continue
             file_page = FilePage(s, "File:" + file_page_title)
             local_path = audio_root / lang.audio_dir_name / v.file[lang.code]
+            temp_wiki_file = temp_download_dir / file_page_title
             if file_page_title in existing:
                 # FIXME: extend this comparison to non-English lines as well?
                 if not file_page_title.startswith("EN"):
                     continue
-                temp_wiki_file = temp_download_dir / file_page_title
+                if not local_path.exists():
+                    # This file is probably removed from the game, but the wiki still has a copy.
+                    continue
                 if not temp_wiki_file.exists():
                     download_file(file_page.get_file_url(), temp_wiki_file)
                 is_same = audio_is_same(local_path, temp_wiki_file)
                 if not is_same:
                     if dry_run:
-                        print(f"{file_page_title} is {is_same}")
+                        print(f"{file_page_title} is different from local copy")
                     elif force_replace:
                         # Only replace the old copy if this is not a dry run AND force replace is explicitly enabled
                         # Need to invalidate the local cache
                         temp_wiki_file.unlink()
-                        upload_audio(local_path, file_page, text, True)
+                        upload_audio(local_path, file_page, text, force=True, temp_wiki_file=temp_wiki_file)
             else:
                 assert local_path.exists()
                 if dry_run:
                     print(f"Will upload {local_path.name} to {file_page_title}")
                 else:
-                    upload_audio(local_path, file_page, text, force=force_replace)
+                    upload_audio(local_path, file_page, text, force=force_replace, temp_wiki_file=temp_wiki_file)
 
 
 def ensure_audio_files_exist(voices: list[Voice]):
